+++ python_modules/dagster-graphql/dagster_graphql/schema/roots/query.py
@@ -14,14 +14,13 @@ from dagster._core.definitions.selector import (
     ScheduleSelector,
     SensorSelector,
 )
-from dagster._core.errors import DagsterCodeLocationNotFoundError, DagsterInvariantViolationError
+from dagster._core.errors import DagsterInvariantViolationError
 from dagster._core.execution.backfill import BulkActionStatus
 from dagster._core.nux import get_has_seen_nux
 from dagster._core.remote_representation.external import CompoundID
 from dagster._core.scheduler.instigation import InstigatorStatus, InstigatorType
 from dagster._core.storage.event_log.base import AssetRecord
 from dagster._core.workspace.permissions import Permissions
-from dagster.components.core.load_defs import PLUGIN_COMPONENT_TYPES_JSON_METADATA_KEY
 
 from dagster_graphql.implementation.execution.backfill import get_asset_backfill_preview
 from dagster_graphql.implementation.external import (
@@ -119,11 +118,7 @@ from dagster_graphql.schema.backfill import (
     GraphenePartitionBackfillsOrError,
 )
 from dagster_graphql.schema.entity_key import GrapheneAssetKey
-from dagster_graphql.schema.env_vars import (
-    GrapheneEnvVarWithConsumersListOrError,
-    GrapheneLocationDocsJson,
-    GrapheneLocationDocsJsonOrError,
-)
+from dagster_graphql.schema.env_vars import GrapheneEnvVarWithConsumersListOrError
 from dagster_graphql.schema.external import (
     GrapheneRepositoriesOrError,
     GrapheneRepositoryConnection,
@@ -315,18 +310,6 @@ class GrapheneQuery(graphene.ObjectType):
         graphene.NonNull(GrapheneEnvVarWithConsumersListOrError),
         repositorySelector=graphene.NonNull(GrapheneRepositorySelector),
         description="Retrieve all the utilized environment variables for the given repo.",
-    )
-
-    hasLocationDocs = graphene.Field(
-        graphene.NonNull(graphene.Boolean),
-        repositorySelector=graphene.NonNull(GrapheneRepositorySelector),
-        description="Retrieves whether the code location has integrated docs.",
-    )
-
-    locationDocsJsonOrError = graphene.Field(
-        graphene.NonNull(GrapheneLocationDocsJsonOrError),
-        repositorySelector=graphene.NonNull(GrapheneRepositorySelector),
-        description="Retrieves JSON blob to drive integrated code location docs.",
     )
 
     sensorOrError = graphene.Field(
@@ -783,41 +766,6 @@ class GrapheneQuery(graphene.ObjectType):
             graphene_info,
             RepositorySelector.from_graphql_input(kwargs.get("repositorySelector")),
         )
-
-    def resolve_hasLocationDocs(
-        self, graphene_info: ResolveInfo, repositorySelector: GrapheneRepositorySelector
-    ):
-        repo_selector = RepositorySelector.from_graphql_input(repositorySelector)
-        try:
-            location = graphene_info.context.get_code_location(repo_selector.location_name)
-            repository = location.get_repository(repo_selector.repository_name)
-        except (DagsterCodeLocationNotFoundError, KeyError):
-            return False
-        return bool(
-            repository.repository_snap.metadata
-            and repository.repository_snap.metadata.get(PLUGIN_COMPONENT_TYPES_JSON_METADATA_KEY)
-        )
-
-    @capture_error
-    def resolve_locationDocsJsonOrError(
-        self, graphene_info: ResolveInfo, repositorySelector: GrapheneRepositorySelector
-    ) -> GrapheneLocationDocsJson:
-        repo_selector = RepositorySelector.from_graphql_input(repositorySelector)
-
-        location = graphene_info.context.get_code_location(repo_selector.location_name)
-        repository = location.get_repository(repo_selector.repository_name)
-        plugin_docs_json = (
-            cast(
-                list,
-                repository.repository_snap.metadata.get(
-                    PLUGIN_COMPONENT_TYPES_JSON_METADATA_KEY, [[]]
-                ),
-            )[0]
-            if repository.repository_snap.metadata
-            else []
-        )
-
-        return GrapheneLocationDocsJson(json=plugin_docs_json)
 
     @capture_error
     def resolve_sensorOrError(
